20231107 backend address validation (#28)
* Add distance_check function that uses Google API.

The function should be used to verify that a users address is close enough to where the robot starts (ORIGIN).

* Ensure that address is close enough to origin.

When users sign up they will not provide an address. This address will be set through the updateUser route after creation. The address must be parseable to the Google API and within a certain radius of the origin.

* Fix purchase_product_order() function in db.py by accepting the same argument as insert_order() and fixing spelling within SQL statement.